+++ frontend/src/components/Datasets.vue
@@ -14,6 +14,20 @@
             </v-col>
         </v-row>
 
+        <v-row>
+            <v-col cols=4>Filter By:</v-col>
+            <v-col cols=8>
+                <Select
+                    :editing="true"
+                    name="ministry_organization"
+                    :label="$tc('Ministry / Organization')"
+                    :items="filterOrgItems"
+                    :value="filterOrg"
+                    @edited=" (newVal) => { filterOrg = newVal; }">
+                </Select>
+            </v-col>
+        </v-row>
+
         <v-list three-line>
             <template v-for="(item, index) in datasetDisplayItems">
                 <v-divider
@@ -45,8 +59,13 @@
 <script>
 
 import {mapActions, mapState, mapMutations} from "vuex";
+import Select from './Select';
 
 export default {
+    components:{
+        Select,
+    },
+
     async created() {
         this.filterBy = this.selectedFilterBy;
 
@@ -55,6 +74,7 @@ export default {
         return {
             message: '',
             filterBy: null,
+            filterOrg: false,
             isUpdating: false,
         }
     },
@@ -83,19 +103,35 @@ export default {
             repos: state => state.repos.repos,
             selectedFilterBy: state => state.repos.selectedFilterBy,
         }),
+
+        filterOrgItems: function(){
+            let rv = [{value: false, text: ""}];
+            let rvValues = [];
+            this.repos.forEach( (repo) => {
+                if (rvValues.indexOf(repo.ministry_organization) === -1){
+                    if (repo.ministry_organization && repo.ministry_organization !== ''){
+                        rvValues.push(repo.ministry_organization);
+                        rv.push({value: repo.ministry_organization, text: repo.ministry_organization});
+                    }
+                }
+            });
+            return rv;
+        },
         
         datasetDisplayItems: function(){
             let items = [];
             this.repos.forEach( (repo, index) => {
-                const item = {
-                    title: `${repo.name}`,
-                    subtitle: repo.create_date,
-                    id: repo._id,
-                    branches: repo.branches,
-                };
-                items.push(item);
-                if(index <= this.repos.length - 1) {
-                    items.push({ divider: true, inset: true });
+                if (!this.filterOrg || repo.ministry_organization===this.filterOrg){
+                    const item = {
+                        title: `${repo.name} ${repo.ministry_organization ? ' - ' + repo.ministry_organization : ''}`,
+                        subtitle: repo.create_date,
+                        id: repo._id,
+                        branches: repo.branches,
+                    };
+                    items.push(item);
+                    if(index <= this.repos.length - 1) {
+                        items.push({ divider: true, inset: true });
+                    }
                 }
             });
             return items;